# openapi.yaml
openapi: "3.1.0"
info:
  title: Qovira API
  version: "0.1"
  description: >-
    The Qovira server API. All endpoints are under /api/v1 and speak application/json. Error responses use application/problem+json (RFC 9457) with a `code` extension field that clients branch on.

    Authentication: session cookie __Host-qovira_session (set by POST /auth/login). CSRF: POST/PATCH/DELETE require a CSRF-Token header echoed from the qovira_csrf cookie (double-submit pattern).
servers:
  - url: /api/v1
    description: Qovira v1 API (same-origin)

# ---------------------------------------------------------------------------
# Reusable schemas
# ---------------------------------------------------------------------------
components:
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, detail, code, requestId]
      properties:
        type:
          type: string
          format: uri
          description: Stable URL identifying the problem kind. Always "https://qovira.ai/errors/" + code.
          example: "https://qovira.ai/errors/not_found"
        title:
          type: string
          description: Short, stable, human-readable summary of the problem type.
          example: "Not found"
        status:
          type: integer
          description: HTTP status code; matches the response status line.
          example: 404
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
          example: "The requested resource was not found."
        code:
          type: string
          description: Snake_case machine slug; the primary client branching key.
          example: "not_found"
        requestId:
          type: string
          description: Correlation ID tying this response to server logs. Always present; sentinel "unknown" if none in context.
          example: "a1b2c3d4e5f6"
        errors:
          type: array
          description: Per-field validation errors. Present on 422 only.
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [pointer, detail]
      properties:
        pointer:
          type: string
          description: RFC 6901 JSON Pointer to the offending field.
          example: "/email"
        detail:
          type: string
          description: Human-readable explanation of why the value is invalid.
          example: "must be a valid email address"
    User:
      type: object
      required: [id, email, displayName, role, timezone, locale, language]
      properties:
        id:
          type: string
          description: Opaque user identifier.
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
        timezone:
          type: string
          description: IANA timezone name (e.g. "America/New_York").
        locale:
          type: string
          description: BCP 47 locale tag (e.g. "en-US").
        language:
          type: string
          description: BCP 47 language tag (e.g. "en").
...
  responses:
    ProblemUnauthorized:
      description: Authentication required (unauthenticated or invalid_credentials).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemForbidden:
      description: CSRF validation failed (csrf_failed).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ProblemUnprocessableEntity:
      description: Validation error (validation_error, validation_failed, invalid_decision).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
paths:
...
  /auth/login:
    post:
      operationId: postAuthLogin
      summary: Log in
      description: Public endpoint. Authenticates with email + password, mints a session, and sets the __Host-qovira_session (HttpOnly) and qovira_csrf (readable) cookies. The session token is never returned in the body.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Authentication successful. Cookies are set.
          content:
            application/json:
              schema:
                type: object
                required: [expiresAt, user]
                properties:
                  expiresAt:
                    type: string
                    format: date-time
                    description: RFC 3339 session expiry timestamp.
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/ProblemUnauthorized"
        "422":
          $ref: "#/components/responses/ProblemUnprocessableEntity"
  /auth/session:
    delete:
      operationId: deleteAuthSession
      summary: Log out (current session)
      description: Invalidates the current session and clears both auth cookies.
      tags: [auth]
      responses:
        "204":
          description: Session cleared.
        "401":
          $ref: "#/components/responses/ProblemUnauthorized"
        "403":
          $ref: "#/components/responses/ProblemForbidden"
  /auth/sessions:
    delete:
      operationId: deleteAuthSessions
      summary: Log out (all sessions)
      description: Invalidates all sessions for the current user and clears auth cookies.
      tags: [auth]
      responses:
        "204":
          description: All sessions cleared.
        "401":
          $ref: "#/components/responses/ProblemUnauthorized"
        "403":
          $ref: "#/components/responses/ProblemForbidden"

  # ---------- Me ----------
  /me:
    get:
      operationId: getMe
      summary: Get current user
      tags: [me]
      responses:
        "200":
          description: Current authenticated user.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/ProblemUnauthorized"

    patch:
      operationId: patchMe
      summary: Update current user profile
      description: "Merge-patch semantics: present fields are set, absent fields unchanged. null is not accepted (use an explicit empty string if you need to clear)."
      tags: [me]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                timezone:
                  type: string
                locale:
                  type: string
                language:
                  type: string
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/ProblemUnauthorized"
        "403":
          $ref: "#/components/responses/ProblemForbidden"
        "422":
          $ref: "#/components/responses/ProblemUnprocessableEntity"
  /me/password:
    post:
      operationId: postMePassword
      summary: Change password
      tags: [me]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        "204":
          description: Password changed successfully.
        "401":
          $ref: "#/components/responses/ProblemUnauthorized"
        "403":
          $ref: "#/components/responses/ProblemForbidden"
        "422":
          $ref: "#/components/responses/ProblemUnprocessableEntity"

  # ---------- Conversations ----------
...